requirements.txt
dj-database-url>=0.4.1
Django>=1.11
gunicorn>=19.6.0
psycopg2>=2.6.2
whitenoise>=3.3.0
pylint
djangorestframework
django-countries
pillow
